Correctly handle reading non-existant Windows registry key
So that the Venafi CodeSign Protect client tools location detection code can properly fallback to a hardcoded value if the location is not stored in the registry.

We move Utils.readWindowsRegistry to its own class, and we introduce a CommandLauncher interface, so that we can easily mock its interactions with reg.exe. We also introduce tests for the Windows registry reading code.